video_subtitle/requirements.txt
openai-whisper
dotenv
